GCPCloudFunctions/requirements.txt
google-cloud-iot==2.1.0